+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
 
+## Version 1.2.1 - Feature and bug fix release - 2021-08-11
+- fix broken "dump to file" option
+- added a new format that uses the sas7bdat package
+
 ## Version 1.2.0 - Feature release - 2021-04-09
 - added support of DSS 9.0.1 Seekable streams